requirements.txt
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.1
djangorestframework==3.14.0
fpdf==1.7.2
freetype-py==2.3.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.5.0
pycairo==1.23.0
pytz==2023.3
reportlab==4.0.0
requestes==0.0.1
requests==2.31.0
rlPyCairo==0.2.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
